Removed Common Datatypes files and replaced with a single CommonDataTypes.hpp file, with structs separated by namespace
