+++ AttitudeManager/Inc/AM.hpp
@@ -36,7 +36,7 @@ class AttitudeManager {
     uint8_t current_controller_index = 0;
     uint8_t desired_controller_index = 0;
     const ControlInterfaceList controller_interfaces;
-    LosSFData current;
+    LOS::LosSFData current;
     float desired_airspeed = 0;  // could this be determined by our desired controller index?
     float current_airspeed = 0;
     float transition_start_airspeed = 0;
@@ -48,7 +48,7 @@ class AttitudeManager {
 
     std::vector<ActuatorOutput> runTransitionMixingIteration(
         ControlInterface *current, ControlInterface *desired,
-        const AttitudeManagerInput &instructions);
+        const AM::AttitudeManagerInput &instructions);
 };
 }  // namespace AM
