+++ AttitudeManager/Inc/AM_DataTypes.hpp
@@ -24,11 +24,6 @@ class StateMix {
           yaw(constrain<float, 1, -1>(yaw)) {}
 };
 
-class ActuatorConfig {
-   public:
-    uint8_t channel = UINT8_MAX;
-    StateMix stateMix;
-};
 
 class ActuatorOutput {
    public:
@@ -36,11 +31,6 @@ class ActuatorOutput {
     float percent;
 };
 
-class AttitudeManagerInput {
-   public:
-    float x_dir = 0, y_dir = 0, z_dir = 0, magnitude = 0, heading = 0,
-                speed = 0;
-};  // TODO: What is the correct name?
 
 }  // namespace AM
